+++ db/events.php
@@ -28,5 +28,9 @@ $observers = [
     [
         'eventname' => '\core\event\user_loggedin',
         'callback' => '\local_profilecohort\profilecohort::set_cohorts_from_profile'
+    ],
+    [
+        'eventname' => '\core\event\user_loggedinas',
+        'callback' => '\local_profilecohort\profilecohort::set_cohorts_from_profile_loginas'
     ]
 ];